% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_above_below.R
\docType{class}
\name{abstract_time_above_or_below_limit_metric-class}
\alias{abstract_time_above_or_below_limit_metric-class}
\title{Abstract time above/below limit class.}
\description{
Abstract time above/below limit class.
}
