% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{applyTimeWindow}
\alias{applyTimeWindow}
\title{Apply time window.}
\usage{
applyTimeWindow(x, window, data_time_unit, ...)
}
\arguments{
\item{x}{input data for the calculation, data frame}

\item{window}{time window}

\item{data_time_unit}{time unit of TIME column in data (x argument)}

\item{...}{extra arguments}
}
\value{
updated object
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`applyTimeWindow()` is deprecated in favor of `apply_time_window()`.
}
