% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{campsisnca_output-class}
\alias{campsisnca_output-class}
\alias{campsisnca_output}
\title{The Campsisnca Output Class Union}
\description{
A class union containing standard data frames, tibbles,
and Campsisnca-specific table objects.
}
