% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{campsisnca_tbl-class}
\alias{campsisnca_tbl-class}
\alias{campsisnca_tbl}
\title{Campsisnca table class (see this class as an interface)}
\description{
Campsisnca table class (see this class as an interface)
}
