% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavg.R
\name{cavg_delegate}
\alias{cavg_delegate}
\title{Compute Cavg (C average).}
\usage{
cavg_delegate(x, variable)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}
}
\value{
individual Cavg
}
\description{
Compute Cavg (C average).
}
