% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metrics.R
\name{cmax_delegate}
\alias{cmax_delegate}
\title{Compute Cmax.}
\usage{
cmax_delegate(x, variable)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}
}
\description{
Compute Cmax.
}
