% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metrics.R
\name{cmin_delegate}
\alias{cmin_delegate}
\title{Compute Cmin.}
\usage{
cmin_delegate(x, variable)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}
}
\description{
Compute Cmin.
}
