% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metrics_s4.R
\name{compute}
\alias{compute}
\title{Compute.}
\usage{
compute(object, ...)
}
\arguments{
\item{object}{metric to be computed}

\item{...}{extra arguments}
}
\value{
updated object
}
\description{
Compute.
}
