% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_utils.R
\name{computeNCAMetricSummary}
\alias{computeNCAMetricSummary}
\title{Compute NCA metric summary.}
\usage{
computeNCAMetricSummary(object, strat_vars, quantile_type)
}
\arguments{
\item{object}{NCA metric}

\item{strat_vars}{stratification variable names in data}

\item{quantile_type}{type of quantile}
}
\value{
data frame
}
\description{
Compute NCA metric summary.
}
