% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_utils.R
\name{computeTableSummary}
\alias{computeTableSummary}
\title{Compute table summary.}
\usage{
computeTableSummary(object)
}
\arguments{
\item{object}{NCA metric}
}
\value{
data frame
}
\description{
Compute table summary.
}
