% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_above_below.R
\name{computeTimeAboveBelow}
\alias{computeTimeAboveBelow}
\title{Compute the duration of a line segment above or below a certain limit.}
\usage{
computeTimeAboveBelow(x1, y1, x2, y2, above, strictly)
}
\arguments{
\item{x1}{x-coordinate of the first point}

\item{y1}{y-coordinate of the first point}

\item{x2}{x-coordinate of the second point}

\item{y2}{y-coordinate of the second point}

\item{above}{whether the line segment is above the limit}

\item{strictly}{whether the line segment is strictly above or below the limit}
}
\value{
the duration of the line segment above or below the limit
}
\description{
Compute the duration of a line segment above or below a certain limit.
}
