% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrough.R
\name{ctrough_delegate}
\alias{ctrough_delegate}
\title{Compute Ctrough}
\usage{
ctrough_delegate(x, variable, time)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}

\item{time}{time value to read Ctrough}
}
\description{
Compute Ctrough
}
