% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{defaultRounding}
\alias{defaultRounding}
\title{Default rounding function (3 significant numbers).}
\usage{
defaultRounding(x, ...)
}
\arguments{
\item{x}{values to be rounded}

\item{...}{extra arguments, not used}
}
\description{
Default rounding function (3 significant numbers).
}
