% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding_utils.R
\name{deparse1Line}
\alias{deparse1Line}
\title{Custom deparse function. Works similarly to deparse1.
However, lines are trimmed before being concatenated.}
\usage{
deparse1Line(x)
}
\arguments{
\item{x}{expression to deparse}
}
\value{
a single string
}
\description{
Custom deparse function. Works similarly to deparse1.
However, lines are trimmed before being concatenated.
}
