% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding_utils.R
\name{deparse_one_line}
\alias{deparse_one_line}
\title{Deparse 1 line.}
\usage{
deparse_one_line(x)
}
\arguments{
\item{x}{expression to deparse}
}
\value{
a single string
}
\description{
Deparse 1 line.
}
