% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding_utils.R
\name{deparseDigits}
\alias{deparseDigits}
\title{Custom deparse function. Works similarly to deparse1.
However, lines are trimmed before being concatenated.}
\usage{
deparseDigits(digits)
}
\arguments{
\item{digits}{rounding digits definitions (integer, function, purrr-style lambda function or list of these)}
}
\value{
a character vector, which will be pasted and given to gtsummary
}
\description{
Custom deparse function. Works similarly to deparse1.
However, lines are trimmed before being concatenated.
}
