% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_utils.R
\name{discard_category_column}
\alias{discard_category_column}
\title{Discard category column.}
\usage{
discard_category_column(x, split = "_")
}
\arguments{
\item{x}{summary export}

\item{split}{string use to concatenate the 'stat' and 'category' column}
}
\value{
updated data frame
}
\description{
Discard category column.
}
