% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{discardCategoryColumn}
\alias{discardCategoryColumn}
\title{Discard category column.}
\usage{
discardCategoryColumn(x, split = "_")
}
\arguments{
\item{x}{summary export}

\item{split}{string use to concatenate the 'stat' and 'category' column}
}
\value{
updated object
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`discardCategoryColumn()` is deprecated in favor of `discard_category_column()`.
}
