% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicated_nca_table.R
\name{export,replicated_nca_table,character-method}
\alias{export,replicated_nca_table,character-method}
\title{Export replicated NCA table.}
\usage{
\S4method{export}{replicated_nca_table,character}(object, dest, ...)
}
\arguments{
\item{object}{replicated_nca_table object}

\item{dest}{destination for the summarised table, either "dataframe", "gtsummary" or "gt"}

\item{...}{extra arguments}
}
\value{
a data frame, a gtsummary object or a gt object, depending on the destination
}
\description{
Export replicated NCA table.
}
