% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_utils.R
\name{extract_brace_values}
\alias{extract_brace_values}
\title{Extract values from within braces.}
\usage{
extract_brace_values(x)
}
\arguments{
\item{x}{input string}
}
\value{
a character vector
}
\description{
Extract values from within braces.
}
