% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_utils.R
\name{extractTableInfo}
\alias{extractTableInfo}
\title{Extract gtsummary table data.}
\usage{
extractTableInfo(tbl)
}
\arguments{
\item{tbl}{gtsummary table}
}
\value{
data frame
}
\description{
Extract gtsummary table data.
}
