% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_utils.R
\name{fillMissingCategoricalLevels}
\alias{fillMissingCategoricalLevels}
\title{Fill missing categorical levels when summary statistics are derived across replicates.}
\usage{
fillMissingCategoricalLevels(x, replace_na = TRUE)
}
\arguments{
\item{x}{summary export across replicates}

\item{replace_na}{replace NA's by 0}
}
\description{
Fill missing categorical levels when summary statistics are derived across replicates.
}
