% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_utils.R
\name{fillMissingCategories}
\alias{fillMissingCategories}
\title{Fill missing categories when summary statistics are computed by campsisnca across.}
\usage{
fillMissingCategories(
  x,
  replace_na = TRUE,
  show_all_dichotomous_levels = TRUE,
  preserve_category_column = FALSE
)
}
\arguments{
\item{x}{summary export across replicates}

\item{replace_na}{replace NA's by 0}

\item{show_all_dichotomous_levels}{show all dichotomous levels (TRUE/FALSE) or only keep TRUE's if FALSE}

\item{preserve_category_column}{preserve category column}
}
\description{
Fill missing categories when summary statistics are computed by campsisnca across.
}
