% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{generate_test_data}
\alias{generate_test_data}
\title{Generate fictitious simulation data for campsisnca README and tests.}
\usage{
generate_test_data()
}
\value{
simulation results
}
\description{
Generate fictitious simulation data for campsisnca README and tests.
}
