% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{generateTestData}
\alias{generateTestData}
\title{Generate fictitious simulation data for campsisnca README and tests.}
\usage{
generateTestData()
}
\value{
simulation results
}
\description{
Generate fictitious simulation data for campsisnca README and tests.
}
