% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_utils.R
\name{get_digits_code}
\alias{get_digits_code}
\title{Get digits code for gtsummary.}
\usage{
get_digits_code(table)
}
\arguments{
\item{table}{NCA table}
}
\value{
code
}
\description{
Get digits code for gtsummary.
}
