% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metric.R, R/cmax.R, R/tmax.R,
%   R/cmin.R, R/tmin.R, R/cat.R, R/ctrough.R, R/cavg.R, R/thalf.R,
%   R/theoretical_thalf.R, R/change_from_baseline.R
\name{get_latex_name}
\alias{get_latex_name}
\alias{get_latex_name,nca_metric-method}
\alias{get_latex_name,abstract_max_metric-method}
\alias{get_latex_name,tmax_metric-method}
\alias{get_latex_name,abstract_min_metric-method}
\alias{get_latex_name,tmin_metric-method}
\alias{get_latex_name,abstract_value_at_metric-method}
\alias{get_latex_name,last_metric-method}
\alias{get_latex_name,ctrough_metric-method}
\alias{get_latex_name,abstract_avg_metric-method}
\alias{get_latex_name,thalf_metric-method}
\alias{get_latex_name,theoretical_thalf_metric-method}
\alias{get_latex_name,cfb_metric-method}
\title{Get the name of the metric in LaTeX notation (with subscript coded with an underscore and brackets).}
\usage{
get_latex_name(x, ...)

\S4method{get_latex_name}{nca_metric}(x)

\S4method{get_latex_name}{abstract_max_metric}(x)

\S4method{get_latex_name}{tmax_metric}(x)

\S4method{get_latex_name}{abstract_min_metric}(x)

\S4method{get_latex_name}{tmin_metric}(x)

\S4method{get_latex_name}{abstract_value_at_metric}(x)

\S4method{get_latex_name}{last_metric}(x)

\S4method{get_latex_name}{ctrough_metric}(x)

\S4method{get_latex_name}{abstract_avg_metric}(x)

\S4method{get_latex_name}{thalf_metric}(x)

\S4method{get_latex_name}{theoretical_thalf_metric}(x)

\S4method{get_latex_name}{cfb_metric}(x)
}
\arguments{
\item{x}{metric}

\item{...}{extra arguments, not used}
}
\value{
a character string
}
\description{
Get the name of the metric in LaTeX notation (with subscript coded with an underscore and brackets).
}
