% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_utils.R
\name{get_statistics_code}
\alias{get_statistics_code}
\title{Get statistics code for gtsummary.}
\usage{
get_statistics_code(table)
}
\arguments{
\item{table}{NCA table}
}
\value{
code
}
\description{
Get statistics code for gtsummary.
}
