% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getCampsisncaOption}
\alias{getCampsisncaOption}
\title{Get Campsisnca option logic.}
\usage{
getCampsisncaOption(name, default)
}
\arguments{
\item{name}{option to search}

\item{default}{default value if option not found}
}
\value{
option value
}
\description{
Get Campsisnca option logic.
}
