% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getCampsisncaOptions}
\alias{getCampsisncaOptions}
\title{Get the Campsisnca options (R options).}
\usage{
getCampsisncaOptions()
}
\value{
global options for Campsisnca
}
\description{
Get the Campsisnca options (R options).
}
\keyword{internal}
