% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metric.R, R/auc.R,
%   R/cmax.R, R/tmax.R, R/cmin.R, R/tmin.R, R/cat.R, R/ctrough.R, R/cavg.R,
%   R/time_above_below.R
\name{getDefaultName}
\alias{getDefaultName}
\alias{getDefaultName,nca_metric-method}
\alias{getDefaultName,auc_metric-method}
\alias{getDefaultName,max_metric-method}
\alias{getDefaultName,cmax_metric-method}
\alias{getDefaultName,tmax_metric-method}
\alias{getDefaultName,min_metric-method}
\alias{getDefaultName,cmin_metric-method}
\alias{getDefaultName,tmin_metric-method}
\alias{getDefaultName,value_at_metric-method}
\alias{getDefaultName,conc_at_metric-method}
\alias{getDefaultName,last_metric-method}
\alias{getDefaultName,ctrough_metric-method}
\alias{getDefaultName,avg_metric-method}
\alias{getDefaultName,cavg_metric-method}
\alias{getDefaultName,time_above_limit_metric-method}
\alias{getDefaultName,time_below_limit_metric-method}
\title{Get default name.}
\usage{
getDefaultName(object, ...)

\S4method{getDefaultName}{nca_metric}(object, ...)

\S4method{getDefaultName}{auc_metric}(object, ...)

\S4method{getDefaultName}{max_metric}(object, ...)

\S4method{getDefaultName}{cmax_metric}(object, ...)

\S4method{getDefaultName}{tmax_metric}(object, ...)

\S4method{getDefaultName}{min_metric}(object, ...)

\S4method{getDefaultName}{cmin_metric}(object, ...)

\S4method{getDefaultName}{tmin_metric}(object, ...)

\S4method{getDefaultName}{value_at_metric}(object, ...)

\S4method{getDefaultName}{conc_at_metric}(object, ...)

\S4method{getDefaultName}{last_metric}(object, ...)

\S4method{getDefaultName}{ctrough_metric}(object, ...)

\S4method{getDefaultName}{avg_metric}(object, ...)

\S4method{getDefaultName}{cavg_metric}(object, ...)

\S4method{getDefaultName}{time_above_limit_metric}(object, ...)

\S4method{getDefaultName}{time_below_limit_metric}(object, ...)
}
\arguments{
\item{object}{get default name of this object}

\item{...}{optional extra arguments}
}
\description{
Get default name.
}
