% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_thalf.R
\name{getDefaultTHalfName}
\alias{getDefaultTHalfName}
\title{Get default name based on thalf subtype.}
\usage{
getDefaultTHalfName(subtype)
}
\arguments{
\item{subtype}{thalf subtype (2cpt.dist, 2cpt.z or 2cpt.eff)}
}
\description{
Get default name based on thalf subtype.
}
