% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_utils.R
\name{getDigitsCode}
\alias{getDigitsCode}
\title{Get digits code for gtsummary.}
\usage{
getDigitsCode(table)
}
\arguments{
\item{table}{NCA table}
}
\value{
code
}
\description{
Get digits code for gtsummary.
}
