% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_utils.R
\name{getLabelsCode}
\alias{getLabelsCode}
\title{Get labels code for gtsummary.}
\usage{
getLabelsCode(table, subscripts)
}
\arguments{
\item{table}{NCA table}

\item{subscripts}{use subscripts, logical value}
}
\value{
code
}
\description{
Get labels code for gtsummary.
}
