% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metrics_table.R
\name{getScenarios}
\alias{getScenarios}
\alias{getScenarios,nca_metrics_table-method}
\title{Get all scenarios that were added to the table object.}
\usage{
getScenarios(object, ...)

\S4method{getScenarios}{nca_metrics_table}(object, ...)
}
\arguments{
\item{object}{table object}

\item{...}{extra arguments}
}
\value{
a dataframe with 2 columns name (stratification variable) and value (all level values)
}
\description{
Get all scenarios that were added to the table object.
}
