% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_utils.R
\name{getStatisticsCode}
\alias{getStatisticsCode}
\title{Get statistics code for gtsummary.}
\usage{
getStatisticsCode(table)
}
\arguments{
\item{table}{NCA table}
}
\value{
code
}
\description{
Get statistics code for gtsummary.
}
