% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metrics.R,
%   R/nca_metrics_table.R
\name{getUnit}
\alias{getUnit}
\alias{getUnit,nca_metrics,character-method}
\alias{getUnit,nca_metrics_table,character-method}
\title{Get the unit corresponding to the given metric.}
\usage{
getUnit(object, metric, ...)

\S4method{getUnit}{nca_metrics,character}(object, metric, ...)

\S4method{getUnit}{nca_metrics_table,character}(object, metric, ...)
}
\arguments{
\item{object}{any object that contains units}

\item{metric}{given metric name}

\item{...}{extra arguments, not used}
}
\description{
Get the unit corresponding to the given metric.
}
