% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_utils.R
\name{getVariableTypeCode}
\alias{getVariableTypeCode}
\title{Get the variable type code for gtsummary.}
\usage{
getVariableTypeCode(table, all_dichotomous_levels)
}
\arguments{
\item{table}{NCA table}

\item{all_dichotomous_levels}{show all dichotomous levels (0 and 1) when data is dichotomous}
}
\value{
code
}
\description{
Get the variable type code for gtsummary.
}
