% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/custom_metric.R
\name{i_value_tbl}
\alias{i_value_tbl}
\alias{i_value_tbl,custom_metric,tbl_df-method}
\title{Compute the individual value of an individual.}
\usage{
i_value_tbl(object, data, ...)

\S4method{i_value_tbl}{custom_metric,tbl_df}(object, data)
}
\arguments{
\item{object}{PK metric}

\item{data}{individual data, tibble}

\item{...}{extra arguments}
}
\value{
individual value
}
\description{
Compute the individual value of an individual.
}
