% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metric.R
\name{i_values}
\alias{i_values}
\alias{i_values,nca_metric-method}
\title{Compute the individual values on a simulation output.}
\usage{
i_values(object, x, options, strat_vars, ...)

\S4method{i_values}{nca_metric}(object, x, options, strat_vars, ...)
}
\arguments{
\item{object}{PK metric}

\item{x}{input data for the calculation, data frame}

\item{options}{NCA options}

\item{strat_vars}{stratification variable names}

\item{...}{extra arguments}
}
\value{
individual values
}
\description{
Compute the individual values on a simulation output.
}
