% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{individual_campsisnca_tbl-class}
\alias{individual_campsisnca_tbl-class}
\alias{individual_campsisnca_tbl}
\title{Individual Campsisnca table class}
\description{
Individual Campsisnca table class
}
