% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{individual_wide_campsisnca_tbl-class}
\alias{individual_wide_campsisnca_tbl-class}
\alias{individual_wide_campsisnca_tbl}
\title{Individual (wide format) Campsisnca table class}
\description{
Individual (wide format) Campsisnca table class
}
