% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_utils.R
\name{isCAMPSIS}
\alias{isCAMPSIS}
\title{Say if the given data frame is a CAMPSIS output or a NONMEM dataset .}
\usage{
isCAMPSIS(x)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}
}
\value{
TRUE for CAMPSIS output, FALSE for NONMEM dataset, error otherwise
}
\description{
Say if the given data frame is a CAMPSIS output or a NONMEM dataset .
}
