% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_methods.R
\name{iValue}
\alias{iValue}
\title{Compute the individual value of an individual.}
\usage{
iValue(object, time, value = NULL)
}
\arguments{
\item{object}{PK metric}

\item{time}{time vector, numeric}

\item{value}{value vector, numeric}
}
\value{
individual value
}
\description{
`r lifecycle::badge("deprecated")`
}
\details{
`iValue()` is deprecated in favor of `i_value()`.
}
