% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metric.R
\name{iValues}
\alias{iValues}
\alias{iValues,nca_metric-method}
\title{Compute the individual values on a simulation output.}
\usage{
iValues(object, x, strat_vars, ...)

\S4method{iValues}{nca_metric}(object, x, strat_vars, ...)
}
\arguments{
\item{object}{PK metric}

\item{x}{input data for the calculation, data frame}

\item{strat_vars}{stratification variable names}

\item{...}{extra arguments}
}
\value{
individual values
}
\description{
Compute the individual values on a simulation output.
}
