% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_from_baseline.R
\docType{class}
\name{log_cfb_metric-class}
\alias{log_cfb_metric-class}
\title{Log-transformed Change from Baseline metric class.}
\description{
Log-transformed Change from Baseline metric class.
}
