% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_export.R
\name{makeKable}
\alias{makeKable}
\title{Make kable.}
\usage{
makeKable(metrics_, vgroup = NULL, vsubgroup = NULL)
}
\arguments{
\item{vgroup}{vertical group variable names}

\item{vsubgroup}{vertical subgroup variable names}

\item{metrics}{dataframe, mandatory columns: metric & cell}
}
\description{
Make kable.
}
