% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_metrics.R
\name{metrics.1cpt}
\alias{metrics.1cpt}
\title{Theoretical metrics for 1-cpt model results.}
\usage{
metrics.1cpt(x, map = character(0))
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{map}{character vector used for column mapping, only one key is possible: K}
}
\value{
theoretical metrics
}
\description{
Theoretical metrics for 1-cpt model results.
}
