% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_metrics.R
\name{metrics.common}
\alias{metrics.common}
\title{Pre-processing for metrics.1cpt and metrics.2cpt.}
\usage{
metrics.common(x, map, thalf.1cpt)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{map}{character vector used for column mapping, only one key is possible: K}

\item{thalf.1cpt}{logical value}
}
\value{
theoretical metrics
}
\description{
Pre-processing for metrics.1cpt and metrics.2cpt.
}
