% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metrics.R
\docType{class}
\name{nca_metrics-class}
\alias{nca_metrics-class}
\title{NCA metrics class. See this class as a list of NCA metrics.}
\description{
NCA metrics class. See this class as a list of NCA metrics.
}
