% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metrics.R
\name{NCAMetrics}
\alias{NCAMetrics}
\title{NCA metrics}
\usage{
NCAMetrics(x = NULL, variable = NULL, scenario)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}

\item{scenario}{character vector used to describe the current scenario.
E.g. c(day="Day 1", condition="Fasted)}
}
\description{
NCA metrics
}
