% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_options.R
\name{NCAOptions}
\alias{NCAOptions}
\title{NCA options used for calculation of metrics.}
\usage{
NCAOptions(
  quantile_type = 2L,
  data_time_unit = "hour",
  table_time_unit = "hour"
)
}
\arguments{
\item{quantile_type}{type of quantile to use (see ?quantile), default value in campsisnca is 2 (aligned with gtsummary)}

\item{data_time_unit}{time unit of the data given to 'calculate'}

\item{table_time_unit}{time unit in table (for time-dependent metrics like AUC, Time above and below, etc.)}
}
\value{
an object of class nca_options
}
\description{
NCA options used for calculation of metrics.
}
