% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_time_range.R
\name{NCATimeRange}
\alias{NCATimeRange}
\title{Create a time range object.}
\usage{
NCATimeRange(
  start = 0,
  end = "last",
  time_unit = "hour",
  exclude_start = FALSE,
  exclude_end = FALSE
)
}
\arguments{
\item{start}{start time of the range}

\item{end}{end time of the range, use 'last' to specify the end of the simulation output}

\item{time_unit}{time unit of 'start' and 'end'}

\item{exclude_start}{exclude start time when filtering}

\item{exclude_end}{exclude end time when filtering}
}
\value{
a time range object
}
\description{
Create a time range object.
}
