% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{getDefaultName}
\alias{getDefaultName}
\title{Get default name.}
\usage{
getDefaultName(object, ...)
}
\arguments{
\item{object}{get default name of this object}

\item{...}{optional extra arguments}
}
\description{
Get default name.
}
