% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metric_utils.R
\name{standardise}
\alias{standardise}
\title{Standardise Campsis/NONMEM dataframe for NCA analysis.
Additional checks will also be performed.}
\usage{
standardise(x, variable, strat_vars)
}
\arguments{
\item{x}{Campsis/NONMEM dataframe}

\item{variable}{dependent variable}

\item{strat_vars}{stratification variables in x (e.g. 'SCENARIO')}
}
\value{
standardised data frame
}
\description{
Standardise Campsis/NONMEM dataframe for NCA analysis.
Additional checks will also be performed.
}
\keyword{internal}
