% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metric.R
\name{stat_display_string}
\alias{stat_display_string}
\alias{stat_display_string,nca_metric-method}
\title{Return the evaluated statistics display string. This method was kept for
backward compatibility in the tests. It is recommended to call `export(dest="dataframe", type="summary_pretty")`
on the NCA table instead.}
\usage{
stat_display_string(object, ...)

\S4method{stat_display_string}{nca_metric}(object, ...)
}
\arguments{
\item{object}{PK metric}

\item{...}{extra arguments}
}
\value{
a string, e.g. 100 [45-143]
}
\description{
Return the evaluated statistics display string. This method was kept for
backward compatibility in the tests. It is recommended to call `export(dest="dataframe", type="summary_pretty")`
on the NCA table instead.
}
