% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{summary_pretty_campsisnca_tbl-class}
\alias{summary_pretty_campsisnca_tbl-class}
\alias{summary_pretty_campsisnca_tbl}
\title{Summary (pretty format) Campsisnca table class}
\description{
Summary (pretty format) Campsisnca table class
}
