% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{summary_wide_campsisnca_tbl-class}
\alias{summary_wide_campsisnca_tbl-class}
\alias{summary_wide_campsisnca_tbl}
\title{Summary (wide format) Campsisnca table class}
\description{
Summary (wide format) Campsisnca table class
}
