% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thalf.R
\name{thalf.1cpt.required}
\alias{thalf.1cpt.required}
\title{Thalf metric required columns for a 1-compartment model.}
\usage{
thalf.1cpt.required()
}
\value{
character vector of required column names
}
\description{
Thalf metric required columns for a 1-compartment model.
}
