% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thalf.R
\name{thalf_delegate}
\alias{thalf_delegate}
\title{Compute individual terminal halflife.}
\usage{
thalf_delegate(x, variable)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}
}
\value{
individual halflife
}
\description{
Compute individual terminal halflife.
}
