% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_utils.R
\name{timerange}
\alias{timerange}
\title{Filter CAMPSIS dataset based on min and max time.}
\usage{
timerange(
  x,
  min = 0,
  max = Inf,
  exclmin = FALSE,
  exclmax = FALSE,
  rebase = FALSE
)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{min}{min time}

\item{max}{max time}

\item{exclmin}{exclude min time when filtering}

\item{exclmax}{exclude max time when filtering}

\item{rebase}{rebase first time to origin, logical value, FALSE by default}
}
\value{
dataset subset
}
\description{
Filter CAMPSIS dataset based on min and max time.
}
