% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metrics.R
\name{toCAMPSISDataframe}
\alias{toCAMPSISDataframe}
\title{Standardise input dataframe to CAMPSIS dataframe.}
\usage{
toCAMPSISDataframe(x)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}
}
\value{
dataframe with id and time columns, filtered on the observations
}
\description{
Standardise input dataframe to CAMPSIS dataframe.
}
