% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_trap_rule.R
\name{trap}
\alias{trap}
\title{NOTE: This method has been adapted from the 'qpNCA' package by Huisman,
Jolling, Mehta and Bergsma (2021) <doi:10.32614/CRAN.package.qpNCA>,
following methodology from Rowland and Tozer (2011,
ISBN:978-0-683-07404-8).}
\usage{
trap(x, y, method = 1)
}
\arguments{
\item{x}{x variable, i.e. time}

\item{y}{y variable, i.e. concentration}

\item{method}{method:
* 1: linear up - linear down
* 2: linear up - logarithmic down
* 3: linear before Tmax, logarithmic after Tmax}
}
\value{
area (length-one numeric)
}
\description{
Calculate Area Under the Curve Using Trapezoids.
}
\details{
Calculates AUC using the trapezoidal method. Assumes data represent a single profile.
Despite choice of method, only linear interpolation is used
for areas of intervals beginning or ending with y: 0.
}
