% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_options.R
\name{UndefinedNCAOptions}
\alias{UndefinedNCAOptions}
\title{Undefined NCA options.}
\usage{
UndefinedNCAOptions()
}
\value{
an object of class undefined_nca_options
}
\description{
Undefined NCA options.
}
