% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_time_window.R
\name{UndefinedTimeWindow}
\alias{UndefinedTimeWindow}
\title{Create an undefined time window.}
\usage{
UndefinedTimeWindow()
}
\value{
undefined time window
}
\description{
Create an undefined time window.
}
