% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_time_window.R
\name{applyTimeWindow}
\alias{applyTimeWindow}
\alias{applyTimeWindow,data.frame,nca_time_window-method}
\alias{applyTimeWindow,tbl_df,nca_time_window-method}
\title{Apply time window.}
\usage{
applyTimeWindow(x, window, ...)

\S4method{applyTimeWindow}{data.frame,nca_time_window}(x, window)

\S4method{applyTimeWindow}{tbl_df,nca_time_window}(x, window)
}
\arguments{
\item{x}{input data for the calculation, data frame}

\item{window}{time window}

\item{...}{extra arguments}
}
\value{
updated object
}
\description{
Apply time window.
}
