% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{applyTimeWindow}
\alias{applyTimeWindow}
\title{Apply time window.}
\usage{
applyTimeWindow(x, window, ...)
}
\arguments{
\item{x}{input data for the calculation, data frame}

\item{window}{time window}

\item{...}{extra arguments}
}
\value{
updated object
}
\description{
Apply time window.
}
