% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_time_window.R
\name{applyTimeWindow}
\alias{applyTimeWindow}
\alias{applyTimeWindow,data.frame,nca_time_window,character-method}
\alias{applyTimeWindow,tbl_df,nca_time_window,character-method}
\title{Apply time window.}
\usage{
applyTimeWindow(x, window, data_time_unit, ...)

\S4method{applyTimeWindow}{data.frame,nca_time_window,character}(x, window, data_time_unit)

\S4method{applyTimeWindow}{tbl_df,nca_time_window,character}(x, window, data_time_unit)
}
\arguments{
\item{x}{input data for the calculation, data frame}

\item{window}{time window}

\item{data_time_unit}{time unit of TIME column in data (x argument)}

\item{...}{extra arguments}
}
\value{
updated object
}
\description{
Apply time window.
}
