% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{AUC}
\alias{AUC}
\title{AUC.}
\usage{
AUC(
  variable = NULL,
  window = TimeWindow(),
  method = 1,
  name = NULL,
  unit = NULL,
  stat_display = NULL,
  digits = NULL
)
}
\arguments{
\item{variable}{dependent variable}

\item{window}{time window on which this metric should be computed}

\item{method}{method:
* 1: linear up - linear down
* 2: linear up - logarithmic down
* 3: linear before Tmax, logarithmic after Tmax}

\item{name}{custom metric name (will be exported into table headers)}

\item{unit}{metric unit (will be exported into table headers if provided)}

\item{stat_display}{statistics display, default is '\{median\} [\{p5\}-\{p95\}]' for continuous data or '\{n\} / \{N\} (\{p\}\%)' for categorical data}

\item{digits}{rounding digits definitions (integer, function, purrr-style lambda function or list of these, 1 item per statistic), see README}
}
\description{
AUC.
}
