% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metrics_s4.R
\name{Auc}
\alias{Auc}
\title{Auc.}
\usage{
Auc(x = NULL, variable = NULL, method = 1)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}

\item{method}{method:
* 1: linear up - linear down
* 2: linear up - logarithmic down
* 3: linear before Tmax, logarithmic after Tmax}
}
\description{
Auc.
}
