% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{Auc}
\alias{Auc}
\title{Auc.}
\usage{
Auc(
  x = NULL,
  variable = NULL,
  method = 1,
  name = NULL,
  unit = NULL,
  stat_display = getStatDisplayDefault()
)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}

\item{method}{method:
* 1: linear up - linear down
* 2: linear up - logarithmic down
* 3: linear before Tmax, logarithmic after Tmax}

\item{name}{custom metric name (will be exported into table headers)}

\item{unit}{metric unit (will be exported into table headers if provided)}

\item{stat_display}{statistics display (see package gtsummary)}
}
\description{
Auc.
}
