% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metric.R,
%   R/theoretical_thalf.R, R/nca_analysis.R, R/nca_analyses.R, R/nca_table.R
\name{calculate}
\alias{calculate}
\alias{calculate,nca_metric,campsis_output,nca_options-method}
\alias{calculate,theoretical_thalf_metric,campsis_output,nca_options-method}
\alias{calculate,nca_analysis,campsis_output,nca_options-method}
\alias{calculate,nca_analyses,campsis_output,nca_options-method}
\alias{calculate,nca_table,campsis_output,nca_options-method}
\title{Calculate.}
\usage{
calculate(object, x, options = NULL, ...)

\S4method{calculate}{nca_metric,campsis_output,nca_options}(object, x, options = NULL, ...)

\S4method{calculate}{theoretical_thalf_metric,campsis_output,nca_options}(object, x, options = NULL, ...)

\S4method{calculate}{nca_analysis,campsis_output,nca_options}(object, x, options = NULL, ...)

\S4method{calculate}{nca_analyses,campsis_output,nca_options}(object, x, options = NULL, ...)

\S4method{calculate}{nca_table,campsis_output,nca_options}(object, x, options = NULL, ...)
}
\arguments{
\item{object}{object (NCA table, NCA analyses, NCA analysis, PK metric) where calculation is applied}

\item{x}{input data for the calculation, data frame}

\item{options}{NCA options}

\item{...}{extra arguments}
}
\value{
updated object
}
\description{
Calculate.
}
