% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metric.R,
%   R/theoretical_thalf.R, R/nca_analysis.R, R/nca_analyses.R,
%   R/nca_metrics_table.R
\name{calculate}
\alias{calculate}
\alias{calculate,nca_metric,data.frame,numeric-method}
\alias{calculate,theoretical_thalf_metric,data.frame,numeric-method}
\alias{calculate,nca_analysis,data.frame,numeric-method}
\alias{calculate,nca_analyses,data.frame,numeric-method}
\alias{calculate,nca_metrics_table,data.frame,numeric-method}
\title{Calculate.}
\usage{
calculate(object, x, quantile_type = NULL, ...)

\S4method{calculate}{nca_metric,data.frame,numeric}(object, x, quantile_type = NULL, ...)

\S4method{calculate}{theoretical_thalf_metric,data.frame,numeric}(object, x, quantile_type = NULL, ...)

\S4method{calculate}{nca_analysis,data.frame,numeric}(object, x, quantile_type = NULL, ...)

\S4method{calculate}{nca_analyses,data.frame,numeric}(object, x, quantile_type = NULL, ...)

\S4method{calculate}{nca_metrics_table,data.frame,numeric}(object, x, quantile_type = NULL, ...)
}
\arguments{
\item{object}{object (PK metric) that needs to be calculated}

\item{x}{input data for the calculation, data frame}

\item{quantile_type}{type of quantile to use (see ?quantile), default value in campsisnca is 2 (aligned with gtsummary)}

\item{...}{extra arguments}
}
\value{
updated object
}
\description{
Calculate.
}
