% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metric.R, R/nca_metrics.R,
%   R/theoretical_thalf.R
\name{calculate}
\alias{calculate}
\alias{calculate,nca_metric,numeric-method}
\alias{calculate,nca_metrics,numeric-method}
\alias{calculate,theoretical_thalf_metric,numeric-method}
\title{Calculate.}
\usage{
calculate(object, quantile_type = NULL, ...)

\S4method{calculate}{nca_metric,numeric}(object, quantile_type = NULL, ...)

\S4method{calculate}{nca_metrics,numeric}(object, quantile_type = NULL, ...)

\S4method{calculate}{theoretical_thalf_metric,numeric}(object, quantile_type = NULL, ...)
}
\arguments{
\item{object}{object (PK metric) that needs to be calculated}

\item{quantile_type}{type of quantile to use (see ?quantile), default value in campsisnca is 2 (aligned with gtsummary)}

\item{...}{extra arguments}
}
\value{
updated object
}
\description{
Calculate.
}
