% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metric.R, R/nca_metrics.R,
%   R/thalf.R
\name{calculate}
\alias{calculate}
\alias{calculate,nca_metric,numeric-method}
\alias{calculate,nca_metrics,numeric-method}
\alias{calculate,thalf_metric,numeric-method}
\title{Calculate.}
\usage{
calculate(object, level = NULL, ...)

\S4method{calculate}{nca_metric,numeric}(object, level = NULL, ...)

\S4method{calculate}{nca_metrics,numeric}(object, level = 0.9, ...)

\S4method{calculate}{thalf_metric,numeric}(object, level = NULL, ...)
}
\arguments{
\item{object}{object (PK metric) that needs to be calculated}

\item{level}{prediction interval level, default is 0.9 (90\% prediction interval)}

\item{...}{extra arguments}
}
\value{
updated object
}
\description{
Calculate.
}
