% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metrics.R, R/auc.R, R/cmax.R,
%   R/tmax.R, R/cmin.R, R/tmin.R, R/ctrough.R, R/cavg.R
\name{calculate}
\alias{calculate}
\alias{calculate,nca_metrics-method}
\alias{calculate,auc_metric-method}
\alias{calculate,cmax_metric-method}
\alias{calculate,tmax_metric-method}
\alias{calculate,cmin_metric-method}
\alias{calculate,tmin_metric-method}
\alias{calculate,ctrough_metric-method}
\alias{calculate,cavg_metric-method}
\title{Calculate.}
\usage{
calculate(object, level = 0.9, ...)

\S4method{calculate}{nca_metrics}(object, level = 0.9, ...)

\S4method{calculate}{auc_metric}(object, level = 0.9, ...)

\S4method{calculate}{cmax_metric}(object, level = 0.9, ...)

\S4method{calculate}{tmax_metric}(object, level = 0.9, ...)

\S4method{calculate}{cmin_metric}(object, level = 0.9, ...)

\S4method{calculate}{tmin_metric}(object, level = 0.9, ...)

\S4method{calculate}{ctrough_metric}(object, level = 0.9, ...)

\S4method{calculate}{cavg_metric}(object, level = 0.9, ...)
}
\arguments{
\item{object}{object (PK metric) that needs to be calculated}

\item{level}{prediction interval level, default is 0.9 (90% prediction interval)}

\item{...}{extra arguments}
}
\value{
updated object
}
\description{
Calculate.
}
