% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{campsis_output-class}
\alias{campsis_output-class}
\alias{campsis_output}
\title{The Campsis Output Class Union}
\description{
A class union containing standard data frames, tibbles,
and Campsis-specific table objects.
}
