% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavg.R
\name{Cavg}
\alias{Cavg}
\title{Cavg.}
\usage{
Cavg(x = NULL, variable = NULL, name = NULL)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}

\item{name}{custom metric name}
}
\description{
Cavg.
}
