% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metrics.R
\name{cavg}
\alias{cavg}
\title{Compute Cavg (C average).}
\usage{
cavg(x, variable)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}
}
\description{
Compute Cavg (C average).
}
