% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavg.R
\name{Cavg}
\alias{Cavg}
\title{Cavg.}
\usage{
Cavg(x = NULL, variable = NULL, name = NULL, unit = NULL)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}

\item{name}{custom metric name (will be exported into table headers)}

\item{unit}{metric unit (will be exported into table headers if provided)}
}
\description{
Cavg.
}
