% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_from_baseline.R
\docType{class}
\name{cfb_metric-class}
\alias{cfb_metric-class}
\title{Change from Baseline metric class.}
\description{
Change from Baseline metric class.
}
\section{Slots}{

\describe{
\item{\code{method}}{Character string specifying the CFB method ("absolute", "percent", "ratio", "log").}
}}

