% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_from_baseline.R
\name{ChangeFromBaseline}
\alias{ChangeFromBaseline}
\title{Change from Baseline (CFB).}
\usage{
ChangeFromBaseline(
  variable = NULL,
  window = NULL,
  name = NULL,
  unit = NULL,
  stat_display = NULL,
  digits = NULL,
  method = "difference"
)
}
\arguments{
\item{variable}{dependent variable}

\item{window}{time window on which this metric should be computed}

\item{name}{custom metric name (will be exported into table headers)}

\item{unit}{metric unit (will be exported into table headers if provided)}

\item{stat_display}{statistics display, default is '\{median\} [\{p5\}-\{p95\}]' for continuous data or '\{n\} / \{N\} (\{p\}\%)' for categorical data}

\item{digits}{rounding digits definitions (integer, function, purrr-style lambda function or list of these, 1 item per statistic), see README}

\item{method}{Character string specifying the calculation method. Must be one of
"difference" (default), "percent", "ratio", or "log".}
}
\description{
Change from Baseline (CFB).
}
