% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmax.R
\name{Cmax}
\alias{Cmax}
\title{Cmax.}
\usage{
Cmax(x = NULL, variable = NULL, name = NULL)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}

\item{name}{custom metric name}
}
\description{
Cmax.
}
