% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmin.R
\name{Cmin}
\alias{Cmin}
\title{Cmin.}
\usage{
Cmin(x = NULL, variable = NULL, name = NULL)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}

\item{name}{custom metric name}
}
\description{
Cmin.
}
