% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_utils.R
\name{computeNCAMetricSummary}
\alias{computeNCAMetricSummary}
\title{Compute NCA metric summary.}
\usage{
computeNCAMetricSummary(object)
}
\arguments{
\item{object}{NCA metric}
}
\value{
data frame
}
\description{
Compute NCA metric summary.
}
