% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_utils.R
\name{computeTableSummary}
\alias{computeTableSummary}
\title{Compute table summary.}
\usage{
computeTableSummary(idata, stat_display)
}
\arguments{
\item{idata}{individual values for a given metric}

\item{stat_display}{statistics display for gtsummary}
}
\value{
data frame
}
\description{
Compute table summary.
}
