% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrough.R
\name{Ctrough}
\alias{Ctrough}
\title{Ctrough.}
\usage{
Ctrough(
  x = NULL,
  variable = NULL,
  time = NULL,
  name = NULL,
  unit = NULL,
  stat_display = NULL,
  digits = NULL
)
}
\arguments{
\item{x}{Campsis/NONMEM dataframe}

\item{variable}{dependent variable}

\item{time}{time value to read Ctrough. If not provided, last concentrations from x will be returned.}

\item{name}{custom metric name (will be exported into table headers)}

\item{unit}{metric unit (will be exported into table headers if provided)}

\item{stat_display}{statistics display, default is '\{median\} [\{p5\}-\{p95\}]' for continuous data or '{n} / {N} ({p}\%)' for categorical data}

\item{digits}{rounding digits definitions (integer, function, purrr-style lambda function or list of these, 1 item per statistic), see README}
}
\description{
Ctrough.
}
