% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctrough.R
\name{Ctrough}
\alias{Ctrough}
\title{Ctrough.}
\usage{
Ctrough(x = NULL, variable = NULL, time, name = NULL, unit = NULL)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}

\item{time}{time value to read Ctrough}

\item{name}{custom metric name (will be exported into table headers)}

\item{unit}{metric unit (will be exported into table headers if provided)}
}
\description{
Ctrough.
}
