% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metrics_s4.R
\name{Ctrough}
\alias{Ctrough}
\title{Ctrough.}
\usage{
Ctrough(x = NULL, variable = NULL, time)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}

\item{time}{time value to read Ctrough}
}
\description{
Ctrough.
}
