% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metrics.R
\name{ctrough}
\alias{ctrough}
\title{Compute Ctrough}
\usage{
ctrough(x, variable, t)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}

\item{t}{time value to read Ctrough}
}
\description{
Compute Ctrough
}
