% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_metric.R
\name{CustomMetric}
\alias{CustomMetric}
\title{Custom metric.}
\usage{
CustomMetric(
  x = NULL,
  variable = NULL,
  fun = function(time, value) {
     0
 },
  name = NULL,
  unit = NULL
)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}

\item{fun}{any custom function with exactly 2 arguments: time and value}

\item{name}{custom metric name (will be exported into table headers)}

\item{unit}{metric unit (will be exported into table headers if provided)}
}
\description{
Custom metric.
}
