% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_metric.R
\name{CustomMetricTbl}
\alias{CustomMetricTbl}
\title{Custom metric (input data as tibble).}
\usage{
CustomMetricTbl(
  window = NULL,
  fun,
  name = NULL,
  unit = NULL,
  categorical = FALSE,
  stat_display = get_stat_display_default(categorical),
  digits = NULL
)
}
\arguments{
\item{window}{time window on which this metric should be computed}

\item{fun}{any custom function with exactly 1 argument: data}

\item{name}{custom metric name (will be exported into table headers)}

\item{unit}{metric unit (will be exported into table headers if provided)}

\item{categorical}{categorical endpoint, logical}

\item{stat_display}{statistics display, default is '\{median\} [\{p5\}-\{p95\}]' for continuous data or '\{n\} / \{N\} (\{p\}\%)' for categorical data}

\item{digits}{rounding digits definitions (integer, function, purrr-style lambda function or list of these, 1 item per statistic), see README}
}
\description{
Custom metric (input data as tibble).
}
