% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_utils.R
\name{discardCategoryColumn}
\alias{discardCategoryColumn}
\title{Discard category column.}
\usage{
discardCategoryColumn(x, split = "_")
}
\arguments{
\item{x}{summary export}

\item{split}{string use to concatenate the 'stat' and 'category' column}
}
\description{
Discard category column.
}
