% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metrics_table.R
\name{.is_replicated}
\alias{.is_replicated}
\title{Does the data contain more than one replicate?}
\usage{
.is_replicated(x)
}
\arguments{
\item{x}{a data frame (typically \code{std_campsis_tbl})}
}
\value{
\code{TRUE} if the data contains a \code{replicate} column with more than one distinct value, \code{FALSE} otherwise
}
\description{
Does the data contain more than one replicate?
}
\keyword{internal}
