% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_utils.R
\name{extractBraceValues}
\alias{extractBraceValues}
\title{Extract values from within braces.}
\usage{
extractBraceValues(x)
}
\arguments{
\item{x}{input string}
}
\value{
a character vector
}
\description{
Extract values from within braces.
}
