% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metrics_table.R
\name{generate_table_code}
\alias{generate_table_code}
\alias{generate_table_code,nca_table,logical-method}
\title{Generate table code.}
\usage{
generate_table_code(object, init = NULL, ...)

\S4method{generate_table_code}{nca_table,logical}(object, init = NULL, ...)
}
\arguments{
\item{object}{table object}

\item{init}{generate initialization code to generate the individuals, default is TRUE}

\item{...}{extra arguments}
}
\description{
Generate table code.
}
