% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metrics_table.R
\name{generateTableCode}
\alias{generateTableCode}
\alias{generateTableCode,nca_metrics_table,logical-method}
\title{Generate table code.}
\usage{
generateTableCode(object, init = NULL, ...)

\S4method{generateTableCode}{nca_metrics_table,logical}(object, init = NULL, ...)
}
\arguments{
\item{object}{table object}

\item{init}{generate initialization code to generate the individuals, default is TRUE}

\item{...}{extra arguments}
}
\description{
Generate table code.
}
