% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metrics_table.R
\name{generateTableCode}
\alias{generateTableCode}
\alias{generateTableCode,nca_metrics_table-method}
\title{Generate table code.}
\usage{
generateTableCode(object, ...)

\S4method{generateTableCode}{nca_metrics_table}(object, ...)
}
\arguments{
\item{object}{table object}

\item{...}{extra arguments}
}
\description{
Generate table code.
}
