% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metric.R, R/auc.R, R/cmax.R,
%   R/tmax.R, R/cmin.R, R/tmin.R, R/cat.R, R/ctrough.R, R/cavg.R, R/thalf.R,
%   R/theoretical_thalf.R, R/time_above_below.R, R/change_from_baseline.R
\name{get_default_name}
\alias{get_default_name}
\alias{get_default_name,nca_metric-method}
\alias{get_default_name,auc_metric-method}
\alias{get_default_name,max_metric-method}
\alias{get_default_name,cmax_metric-method}
\alias{get_default_name,tmax_metric-method}
\alias{get_default_name,min_metric-method}
\alias{get_default_name,cmin_metric-method}
\alias{get_default_name,tmin_metric-method}
\alias{get_default_name,value_at_metric-method}
\alias{get_default_name,conc_at_metric-method}
\alias{get_default_name,last_metric-method}
\alias{get_default_name,ctrough_metric-method}
\alias{get_default_name,avg_metric-method}
\alias{get_default_name,cavg_metric-method}
\alias{get_default_name,thalf_metric-method}
\alias{get_default_name,theoretical_thalf_metric-method}
\alias{get_default_name,time_above_limit_metric-method}
\alias{get_default_name,time_below_limit_metric-method}
\alias{get_default_name,cfb_metric-method}
\title{Get default name.}
\usage{
get_default_name(object, ...)

\S4method{get_default_name}{nca_metric}(object, ...)

\S4method{get_default_name}{auc_metric}(object, ...)

\S4method{get_default_name}{max_metric}(object, ...)

\S4method{get_default_name}{cmax_metric}(object, ...)

\S4method{get_default_name}{tmax_metric}(object, ...)

\S4method{get_default_name}{min_metric}(object, ...)

\S4method{get_default_name}{cmin_metric}(object, ...)

\S4method{get_default_name}{tmin_metric}(object, ...)

\S4method{get_default_name}{value_at_metric}(object, ...)

\S4method{get_default_name}{conc_at_metric}(object, ...)

\S4method{get_default_name}{last_metric}(object, ...)

\S4method{get_default_name}{ctrough_metric}(object, ...)

\S4method{get_default_name}{avg_metric}(object, ...)

\S4method{get_default_name}{cavg_metric}(object, ...)

\S4method{get_default_name}{thalf_metric}(object, ...)

\S4method{get_default_name}{theoretical_thalf_metric}(object, ...)

\S4method{get_default_name}{time_above_limit_metric}(object, ...)

\S4method{get_default_name}{time_below_limit_metric}(object, ...)

\S4method{get_default_name}{cfb_metric}(object, ...)
}
\arguments{
\item{object}{get default name of this object}

\item{...}{optional extra arguments}
}
\description{
Get default name.
}
