% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_thalf.R
\name{get_default_thalf_name}
\alias{get_default_thalf_name}
\title{Get default name based on thalf subtype.}
\usage{
get_default_thalf_name(subtype)
}
\arguments{
\item{subtype}{thalf subtype (2cpt.dist, 2cpt.z or 2cpt.eff)}
}
\description{
Get default name based on thalf subtype.
}
