% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_table.R
\name{get_strata}
\alias{get_strata}
\alias{get_strata,nca_table,logical-method}
\title{Get all stratas.}
\usage{
get_strata(object, keep_single = NULL, ...)

\S4method{get_strata}{nca_table,logical}(object, keep_single = NULL, ...)
}
\arguments{
\item{object}{table object}

\item{keep_single}{keep single stratification values, logical (default TRUE)}

\item{...}{extra arguments}
}
\value{
list of stratification variable names
}
\description{
Get all stratas.
}
