% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metrics.R, R/nca_analysis.R,
%   R/nca_metrics_table.R
\name{get_unit}
\alias{get_unit}
\alias{get_unit,nca_metrics,character-method}
\alias{get_unit,nca_analysis,character-method}
\alias{get_unit,nca_metrics_table,character-method}
\title{Get the unit corresponding to the given metric.}
\usage{
get_unit(object, metric, ...)

\S4method{get_unit}{nca_metrics,character}(object, metric, ...)

\S4method{get_unit}{nca_analysis,character}(object, metric, ...)

\S4method{get_unit}{nca_metrics_table,character}(object, metric, ...)
}
\arguments{
\item{object}{any object that contains units}

\item{metric}{given metric name}

\item{...}{extra arguments, not used}
}
\description{
Get the unit corresponding to the given metric.
}
