% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metric.R, R/cmax.R, R/tmax.R,
%   R/cmin.R, R/tmin.R, R/ctrough.R, R/cavg.R, R/thalf.R, R/theoretical_thalf.R
\name{getLaTeXName}
\alias{getLaTeXName}
\alias{getLaTeXName,nca_metric-method}
\alias{getLaTeXName,tmax_metric-method}
\alias{getLaTeXName,cmin_metric-method}
\alias{getLaTeXName,tmin_metric-method}
\alias{getLaTeXName,ctrough_metric-method}
\alias{getLaTeXName,cavg_metric-method}
\alias{getLaTeXName,thalf_metric-method}
\alias{getLaTeXName,theoretical_thalf_metric-method}
\title{Get the name of the metric in LaTeX notation (with subscript coded with an underscore and brackets).}
\usage{
getLaTeXName(x, ...)

\S4method{getLaTeXName}{nca_metric}(x)

\S4method{getLaTeXName}{nca_metric}(x)

\S4method{getLaTeXName}{tmax_metric}(x)

\S4method{getLaTeXName}{cmin_metric}(x)

\S4method{getLaTeXName}{tmin_metric}(x)

\S4method{getLaTeXName}{ctrough_metric}(x)

\S4method{getLaTeXName}{cavg_metric}(x)

\S4method{getLaTeXName}{thalf_metric}(x)

\S4method{getLaTeXName}{theoretical_thalf_metric}(x)
}
\arguments{
\item{x}{metric}

\item{...}{extra arguments, not used}
}
\description{
Get the name of the metric in LaTeX notation (with subscript coded with an underscore and brackets).
}
