% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_utils.R
\name{getTableSummaryCode}
\alias{getTableSummaryCode}
\title{Get table summary code.}
\usage{
getTableSummaryCode(
  variable,
  data,
  by,
  stats,
  type,
  labels,
  digits,
  combine_with,
  header_label
)
}
\arguments{
\item{variable}{assigned variable name}

\item{data}{data frame code}

\item{by}{variable}

\item{stats}{stats to compute}

\item{type}{type of the variables}

\item{labels}{the labels to display}

\item{digits}{the digits to be used for rounding}

\item{combine_with}{either 'tbl_stack' or 'tbl_merge'}

\item{header_label}{header label name}
}
\value{
data frame
}
\description{
Get table summary code.
}
