% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_utils.R
\name{getTableSummaryCode}
\alias{getTableSummaryCode}
\title{Get table summary code.}
\usage{
getTableSummaryCode(data, by, stats)
}
\arguments{
\item{data}{data frame code}

\item{by}{variable}

\item{stats}{stats to compute}
}
\value{
data frame
}
\description{
Get table summary code.
}
