% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_utils.R
\name{getTableSummaryCode}
\alias{getTableSummaryCode}
\title{Get table summary code.}
\usage{
getTableSummaryCode(variable, data, by, stats, labels)
}
\arguments{
\item{data}{data frame code}

\item{by}{variable}

\item{stats}{stats to compute}

\item{labels}{the labels to display}
}
\value{
data frame
}
\description{
Get table summary code.
}
