% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{getUnit}
\alias{getUnit}
\title{Get the unit corresponding to the given metric.}
\usage{
getUnit(object, metric, ...)
}
\arguments{
\item{object}{any object that contains units}

\item{metric}{given metric name}
}
\description{
Get the unit corresponding to the given metric.
}
