% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_utils.R
\name{getVariableTypeCode}
\alias{getVariableTypeCode}
\title{Get the variable type code for gtsummary.}
\usage{
getVariableTypeCode(table, subscripts)
}
\arguments{
\item{table}{NCA table}
}
\value{
code
}
\description{
Get the variable type code for gtsummary.
}
