% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_utils.R
\name{glueStatDisplay}
\alias{glueStatDisplay}
\title{Glue stat display string.}
\usage{
glueStatDisplay(stat_display, stats, summary)
}
\arguments{
\item{stat_display}{stat display string}

\item{stats}{statistics, character vector}

\item{summary}{summary data frame}
}
\value{
glued string
}
\description{
Glue stat display string.
}
