% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/auc.R, R/cmax.R, R/tmax.R,
%   R/cmin.R, R/tmin.R, R/cat.R, R/ctrough.R, R/cavg.R, R/thalf.R,
%   R/time_above_below.R, R/change_from_baseline.R, R/custom_metric.R
\name{i_value}
\alias{i_value}
\alias{i_value,auc_metric,numeric,numeric-method}
\alias{i_value,abstract_max_metric,numeric,numeric-method}
\alias{i_value,tmax_metric,numeric,numeric-method}
\alias{i_value,abstract_min_metric,numeric,numeric-method}
\alias{i_value,tmin_metric,numeric,numeric-method}
\alias{i_value,abstract_value_at_metric,numeric,numeric-method}
\alias{i_value,abstract_last_metric,numeric,numeric-method}
\alias{i_value,abstract_avg_metric,numeric,numeric-method}
\alias{i_value,thalf_metric,numeric,numeric-method}
\alias{i_value,abstract_time_above_or_below_limit_metric,numeric,numeric-method}
\alias{i_value,baseline_metric,numeric,numeric-method}
\alias{i_value,custom_metric,numeric,numeric-method}
\title{Compute the individual value of an individual.}
\usage{
i_value(object, time, value = NULL)

\S4method{i_value}{auc_metric,numeric,numeric}(object, time, value = NULL)

\S4method{i_value}{abstract_max_metric,numeric,numeric}(object, time, value = NULL)

\S4method{i_value}{tmax_metric,numeric,numeric}(object, time, value = NULL)

\S4method{i_value}{abstract_min_metric,numeric,numeric}(object, time, value = NULL)

\S4method{i_value}{tmin_metric,numeric,numeric}(object, time, value = NULL)

\S4method{i_value}{abstract_value_at_metric,numeric,numeric}(object, time, value = NULL)

\S4method{i_value}{abstract_last_metric,numeric,numeric}(object, time, value = NULL)

\S4method{i_value}{abstract_avg_metric,numeric,numeric}(object, time, value = NULL)

\S4method{i_value}{thalf_metric,numeric,numeric}(object, time, value = NULL)

\S4method{i_value}{abstract_time_above_or_below_limit_metric,numeric,numeric}(object, time, value = NULL)

\S4method{i_value}{baseline_metric,numeric,numeric}(object, time, value = NULL)

\S4method{i_value}{custom_metric,numeric,numeric}(object, time, value = NULL)
}
\arguments{
\item{object}{PK metric}

\item{time}{time vector, numeric}

\item{value}{value vector, numeric}
}
\value{
individual value
}
\description{
Compute the individual value of an individual.
}
