% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/auc.R, R/cmax.R, R/tmax.R,
%   R/cmin.R, R/tmin.R, R/ctrough.R, R/cavg.R, R/thalf.R, R/custom_metric.R
\name{iValue}
\alias{iValue}
\alias{iValue,auc_metric,numeric,numeric-method}
\alias{iValue,cmax_metric,numeric,numeric-method}
\alias{iValue,tmax_metric,numeric,numeric-method}
\alias{iValue,cmin_metric,numeric,numeric-method}
\alias{iValue,tmin_metric,numeric,numeric-method}
\alias{iValue,ctrough_metric,numeric,numeric-method}
\alias{iValue,cavg_metric,numeric,numeric-method}
\alias{iValue,thalf_metric,numeric,numeric-method}
\alias{iValue,custom_metric,numeric,numeric-method}
\title{Compute the individual value of an individual.}
\usage{
iValue(object, time, value, ...)

\S4method{iValue}{auc_metric,numeric,numeric}(object, time, value, method)

\S4method{iValue}{cmax_metric,numeric,numeric}(object, time, value)

\S4method{iValue}{tmax_metric,numeric,numeric}(object, time, value)

\S4method{iValue}{cmin_metric,numeric,numeric}(object, time, value)

\S4method{iValue}{tmin_metric,numeric,numeric}(object, time, value)

\S4method{iValue}{ctrough_metric,numeric,numeric}(object, time, value, trough_time)

\S4method{iValue}{cavg_metric,numeric,numeric}(object, time, value)

\S4method{iValue}{thalf_metric,numeric,numeric}(object, time, value)

\S4method{iValue}{custom_metric,numeric,numeric}(object, time, value, custom_function)
}
\arguments{
\item{object}{PK metric}

\item{time}{time vector, numeric}

\item{value}{value vector, numeric}

\item{...}{extra arguments}

\item{method}{method:
* 1: linear up - linear down
* 2: linear up - logarithmic down
* 3: linear before Tmax, logarithmic after Tmax}

\item{trough_time}{trough sample time, numeric value}

\item{custom_function}{any R function with exactly 2 arguments: time and value}
}
\value{
individual value
}
\description{
Compute the individual value of an individual.
}
