% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/cmin.R
\name{iValue}
\alias{iValue}
\alias{iValue,cmin_metric,numeric,numeric-method}
\title{Compute the individual value of an individual.}
\usage{
iValue(object, time, value, ...)

\S4method{iValue}{cmin_metric,numeric,numeric}(object, time, value)
}
\arguments{
\item{object}{PK metric}

\item{time}{time vector, numeric}

\item{value}{value vector, numeric}

\item{...}{extra arguments}
}
\value{
individual value
}
\description{
Compute the individual value of an individual.
}
