% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metric.R
\name{iValues}
\alias{iValues}
\alias{iValues,nca_metric-method}
\title{Compute the individual values of a population.}
\usage{
iValues(object, ...)

\S4method{iValues}{nca_metric}(object, ...)
}
\arguments{
\item{object}{PK metric}

\item{...}{extra arguments}
}
\value{
individual values
}
\description{
Compute the individual values of a population.
}
