% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_export.R
\name{makeKable}
\alias{makeKable}
\title{Make kable.}
\usage{
makeKable(x, df, vgroup = NULL, vsubgroup = NULL, format = "html")
}
\arguments{
\item{x}{table object}

\item{df}{output of makeTable, dataframe}

\item{vgroup}{vertical group variable names}

\item{vsubgroup}{vertical subgroup variable names}

\item{format}{can be 'html' or "pdf'}
}
\description{
Make kable.
}
