% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_metrics.R
\name{metrics.2cpt}
\alias{metrics.2cpt}
\title{Theoretical metrics for 2-cpt model results.}
\usage{
metrics.2cpt(x, map = character(0))
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{map}{character vector used for column mapping, keys to be chosen among: DOSE, TAU, CL, V2, Q, V3, KA}
}
\value{
theoretical metrics
}
\description{
Theoretical metrics for 2-cpt model results.
}
