% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_analysis.R
\name{NCAAnalysis}
\alias{NCAAnalysis}
\title{Create an NCA analysis.}
\usage{
NCAAnalysis(
  name = "Default",
  window = TimeWindow(),
  variable = NULL,
  strata = get_default_strata()
)
}
\arguments{
\item{name}{name of this analysis, e.g. 'Day 1'}

\item{window}{time window, see \link{TimeWindow}}

\item{variable}{default variable which is analysed}

\item{strata}{strata levels this analysis refers to, named vector, e.g. c(ARM='1g QD').
Note, the default strata are c(SCENARIO='all', ARM='all').
Use 'all' if this analysis refers to all levels for the specified stratification variable.
By default, a stratification variable that has only 1 level is ignored.}
}
\description{
Create an NCA analysis.
}
