% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_analysis.R
\name{NCAAnalysis}
\alias{NCAAnalysis}
\title{Create an NCA analysis.}
\usage{
NCAAnalysis(
  name = "Default",
  window = TimeWindow(),
  variable = NULL,
  strata = NULL
)
}
\arguments{
\item{name}{name of this analysis, e.g. 'Day 1'}

\item{window}{time window, see \link{TimeWindow}}

\item{variable}{default variable which is analysed}

\item{strata}{specific strata this analysis refers to, named vector (e.g. c(ARM='1g QD'))}
}
\description{
Create an NCA analysis.
}
