% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_analysis.R
\name{NCAAnalysis}
\alias{NCAAnalysis}
\title{Create an NCA analysis.}
\usage{
NCAAnalysis(
  name = "Default",
  window = TimeWindow(),
  variable = NULL,
  metrics = NCAMetrics()
)
}
\arguments{
\item{name}{name of this analysis, e.g. 'Day 1'}

\item{variable}{default variable which is analysed}

\item{metrics}{list of metrics}

\item{time_range}{default time range for this analysis}
}
\description{
Create an NCA analysis.
}
