% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metrics.R
\name{NCAMetrics}
\alias{NCAMetrics}
\title{NCA metrics}
\usage{
NCAMetrics(x = NULL, variable = NULL, scenario)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}

\item{time}{time value to read Ctrough}
}
\description{
NCA metrics
}
