% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metrics_table.R
\name{NCAMetricsTable}
\alias{NCAMetricsTable}
\title{NCA metrics table.}
\usage{
NCAMetricsTable(json = NULL)
}
\arguments{
\item{json}{path to JSON table file or JSON content in string form}
}
\description{
NCA metrics table.
}
