% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metrics_table.R
\name{NCAMetricsTable}
\alias{NCAMetricsTable}
\title{NCA metrics table.}
\usage{
NCAMetricsTable(rounding = NULL, unitLineBreak = FALSE)
}
\arguments{
\item{rounding}{rounding function (default arguments should be: x (values), 
metric (metric name), stat (low, med, up))}

\item{unitLineBreak}{line break between header and unit, logical value}
}
\description{
NCA metrics table.
}
