% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_options.R
\name{NCAOptions}
\alias{NCAOptions}
\title{NCA options used for calculation of metrics.}
\usage{
NCAOptions(
  quantile_type = 2L,
  data_time_unit = "hour",
  table_time_unit = "hour",
  rep_stat_display = getStatDisplayDefault(),
  rep_stat_digits = 3L
)
}
\arguments{
\item{quantile_type}{type of quantile to use (see ?quantile), default value in campsisnca is 2 (aligned with gtsummary)}

\item{data_time_unit}{time unit of the data given to 'calculate'}

\item{table_time_unit}{time unit in table (for time-dependent metrics like AUC, Time above and below, etc.)}

\item{rep_stat_display}{display format for replicate statistics, character vector. Default is \verb{'{median} ({p5}–{p95})'}.}

\item{rep_stat_digits}{number of significant digits to display for replicate statistics, default is 3.}
}
\description{
NCA options used for calculation of metrics.
}
