% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metrics_table.R
\name{NCATable}
\alias{NCATable}
\title{NCA table.}
\usage{
NCATable(
  title = NULL,
  subtitle = NULL,
  nca_options = NCAOptions(),
  tab_options = list(),
  json = NULL
)
}
\arguments{
\item{title}{table title, optional character value}

\item{subtitle}{table subtitle, optional character value}

\item{nca_options}{NCA options, see ?NCAOptions}

\item{tab_options}{list of options to pass to gt::tab_options}

\item{json}{path to JSON table file or JSON content in string form}
}
\description{
NCA table.
}
