% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metrics.R,
%   R/nca_metrics_table.R
\name{reduceTo2Dimensions}
\alias{reduceTo2Dimensions}
\alias{reduceTo2Dimensions,nca_metrics-method}
\alias{reduceTo2Dimensions,nca_metrics_table-method}
\title{Reduce table to 2 dimensions.}
\usage{
reduceTo2Dimensions(object, ...)

\S4method{reduceTo2Dimensions}{nca_metrics}(object, ...)

\S4method{reduceTo2Dimensions}{nca_metrics_table}(object, ...)
}
\arguments{
\item{object}{generic object}

\item{...}{extra arguments}
}
\value{
same object/table with max 2 dimensions (useful before export)
}
\description{
Reduce table to 2 dimensions.
}
