% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicated_nca_options.R
\name{ReplicatedNCAOptions}
\alias{ReplicatedNCAOptions}
\title{Replicated NCA options.}
\usage{
ReplicatedNCAOptions(
  selected_stastistics = character(),
  summary_stat_display = get_stat_display_default(),
  summary_stat_signif_digits = 3L
)
}
\arguments{
\item{selected_stastistics}{NCA metrics statistics to keep (e.g. mean, etc) when summary statistics are computed on replicated output.
Default is the empty character vector (all statistics are computed).}

\item{summary_stat_display}{display format for replicate statistics, character vector. Default is \verb{'{median} ({p5}–{p95})'}.}

\item{summary_stat_signif_digits}{number of significant digits to display for replicate statistics, default is 3.}
}
\description{
Replicated NCA options.
}
