% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metrics.R
\name{standardise}
\alias{standardise}
\title{Standardise CAMPSIS dataframe for NCA analysis.
Additional checks will also be performed.}
\usage{
standardise(x, variable)
}
\arguments{
\item{x}{CAMPSIS dataframe}

\item{variable}{dependent variable}
}
\description{
Standardise CAMPSIS dataframe for NCA analysis.
Additional checks will also be performed.
}
