% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metric.R
\name{statDisplayString}
\alias{statDisplayString}
\alias{statDisplayString,nca_metric-method}
\title{Return the evaluated statistics display string.}
\usage{
statDisplayString(object, ...)

\S4method{statDisplayString}{nca_metric}(object, ...)
}
\arguments{
\item{object}{PK metric}

\item{...}{extra arguments}
}
\value{
a string, e.g. 100 [45-143]
}
\description{
Return the evaluated statistics display string.
}
