% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/nca_metrics_table.R
\name{summarise_replicates}
\alias{summarise_replicates}
\alias{summarise_replicates,nca_metrics_table,campsisnca_output,nca_options-method}
\title{Summarise replicates.}
\usage{
summarise_replicates(object, x, options = NULL, ...)

\S4method{summarise_replicates}{nca_metrics_table,campsisnca_output,nca_options}(object, x, options = NULL, ...)
}
\arguments{
\item{object}{object (NCA table) which contains the table structure}

\item{x}{input data for the calculation, data frame of class 'individual_campsisnca_tbl' or 'individual_wide_campsisnca_tbl'}

\item{options}{NCA options}

\item{...}{extra arguments}
}
\value{
updated object
}
\description{
Summarise replicates.
}
