% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thalf.R
\name{Thalf.1cpt}
\alias{Thalf.1cpt}
\title{Theoretical half life for a 1-compartment model.}
\usage{
Thalf.1cpt(x = NULL, map = NULL)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{map}{character vector used for column mapping, only one key is possible: K}
}
\description{
Theoretical half life for a 1-compartment model.
}
