% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_thalf.R
\name{Thalf.2cpt.dist}
\alias{Thalf.2cpt.dist}
\title{Theoretical distribution half life for a 2-compartment model.}
\usage{
Thalf.2cpt.dist(
  x = NULL,
  map = NULL,
  name = NULL,
  unit = NULL,
  stat_display = getStatDisplayDefault(),
  digits = NULL
)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{map}{character vector used for column mapping, keys to be chosen among: DOSE, TAU, CL, V2, Q, V3, KA}

\item{name}{custom metric name (will be exported into table headers)}

\item{unit}{metric unit (will be exported into table headers if provided)}

\item{stat_display}{statistics display (see package gtsummary)}

\item{digits}{rounding digits definitions (integer, function, purrr-style lambda function or list of these, 1 item per statistic), see README}
}
\description{
Theoretical distribution half life for a 2-compartment model.
}
