% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thalf.R
\name{Thalf.2cpt.dist}
\alias{Thalf.2cpt.dist}
\title{Theoretical distribution half life for a 2-compartment model.}
\usage{
Thalf.2cpt.dist(x = NULL, map = NULL, name = NULL, unit = NULL)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{map}{character vector used for column mapping, keys to be chosen among: DOSE, TAU, CL, V2, Q, V3, KA}

\item{name}{custom metric name (will be exported into table headers)}

\item{unit}{metric unit (will be exported into table headers if provided)}
}
\description{
Theoretical distribution half life for a 2-compartment model.
}
