% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_thalf.R
\name{Thalf.2cpt.eff}
\alias{Thalf.2cpt.eff}
\title{Theoretical effective half life for a 2-compartment model.}
\usage{
Thalf.2cpt.eff(
  x = NULL,
  map = NULL,
  name = NULL,
  unit = NULL,
  stat_display = getStatDisplayDefault()
)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{map}{character vector used for column mapping, keys to be chosen among: DOSE, TAU, CL, V2, Q, V3, KA}

\item{name}{custom metric name (will be exported into table headers)}

\item{unit}{metric unit (will be exported into table headers if provided)}

\item{stat_display}{statistics display (see package gtsummary)}
}
\description{
Theoretical effective half life for a 2-compartment model.
}
