% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thalf.R
\name{Thalf.2cpt.z}
\alias{Thalf.2cpt.z}
\title{Theoretical elimination half life for a 2-compartment model.}
\usage{
Thalf.2cpt.z(x = NULL, map = NULL, name = NULL)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{map}{character vector used for column mapping, keys to be chosen among: DOSE, TAU, CL, V2, Q, V3, KA}

\item{name}{custom metric name}
}
\description{
Theoretical elimination half life for a 2-compartment model.
}
