% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thalf.R
\name{Thalf}
\alias{Thalf}
\title{Terminal half life computed by making a linear regression in the log domain
on the given data x.}
\usage{
Thalf(
  variable = NULL,
  window = NULL,
  name = NULL,
  unit = NULL,
  stat_display = NULL,
  digits = NULL
)
}
\arguments{
\item{variable}{dependent variable}

\item{window}{time window on which this metric should be computed}

\item{name}{custom metric name (will be exported into table headers)}

\item{unit}{metric unit (will be exported into table headers if provided)}

\item{stat_display}{statistics display, default is '\{median\} [\{p5\}-\{p95\}]' for continuous data or '\{n\} / \{N\} (\{p\}\%)' for categorical data}

\item{digits}{rounding digits definitions (integer, function, purrr-style lambda function or list of these, 1 item per statistic), see README}
}
\description{
Terminal half life computed by making a linear regression in the log domain
on the given data x.
}
