% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thalf.R
\name{Thalf}
\alias{Thalf}
\title{Terminal half life computed by making a linear regression in the log domain 
on the given data x.}
\usage{
Thalf(
  x = NULL,
  variable = NULL,
  name = NULL,
  unit = NULL,
  stat_display = getStatDisplayDefault()
)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}

\item{name}{custom metric name (will be exported into table headers)}

\item{unit}{metric unit (will be exported into table headers if provided)}

\item{stat_display}{statistics display (see package gtsummary)}
}
\description{
Terminal half life computed by making a linear regression in the log domain 
on the given data x.
}
