% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_above_below.R
\name{TimeBelowLimit}
\alias{TimeBelowLimit}
\title{Time below a certain limit.}
\usage{
TimeBelowLimit(
  x = NULL,
  variable = NULL,
  limit = NULL,
  strictly = FALSE,
  name = NULL,
  unit = NULL,
  stat_display = NULL,
  digits = NULL
)
}
\arguments{
\item{x}{Campsis/NONMEM dataframe}

\item{variable}{dependent variable}

\item{limit}{the limit to compare the variable against}

\item{strictly}{whether the variable must be strictly below the limit}

\item{name}{custom metric name (will be exported into table headers)}

\item{unit}{metric unit (will be exported into table headers if provided)}

\item{stat_display}{statistics display, default is '\{median\} [\{p5\}-\{p95\}]' for continuous data or '\{n\} / \{N\} (\{p\}\%)' for categorical data}

\item{digits}{rounding digits definitions (integer, function, purrr-style lambda function or list of these, 1 item per statistic), see README}
}
\description{
Time below a certain limit.
}
