% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_utilities.R
\name{timerange}
\alias{timerange}
\title{Filter CAMPSIS dataset based on min and max time.}
\usage{
timerange(x, min = 0, max = Inf, exclmin = FALSE, exclmax = FALSE)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{min}{min time}

\item{max}{max time}

\item{exclmin}{exclude min time when filtering}

\item{exclmax}{exclude max time when filtering}
}
\value{
dataset subset
}
\description{
Filter CAMPSIS dataset based on min and max time.
}
