% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmax.R
\name{Tmax}
\alias{Tmax}
\title{Tmax.}
\usage{
Tmax(x = NULL, variable = NULL, name = NULL)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}

\item{name}{custom metric name}
}
\description{
Tmax.
}
