% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metrics.R
\name{tmax_delegate}
\alias{tmax_delegate}
\title{Compute tmax.}
\usage{
tmax_delegate(x, variable)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}
}
\description{
Compute tmax.
}
