% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmin.R
\name{Tmin}
\alias{Tmin}
\title{Tmin.}
\usage{
Tmin(x = NULL, variable = NULL, name = NULL)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}

\item{name}{custom metric name}
}
\description{
Tmin.
}
