% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metrics.R
\name{tmin}
\alias{tmin}
\title{Compute tmin.}
\usage{
tmin(x, variable)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}

\item{after}{delay, when do we start looking at concentrations}
}
\description{
Compute tmin.
}
