% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmin.R
\name{Tmin}
\alias{Tmin}
\title{Tmin.}
\usage{
Tmin(x = NULL, variable = NULL, name = NULL, unit = NULL)
}
\arguments{
\item{x}{CAMPSIS/NONMEM dataframe}

\item{variable}{dependent variable}

\item{name}{custom metric name (will be exported into table headers)}

\item{unit}{metric unit (will be exported into table headers if provided)}
}
\description{
Tmin.
}
