% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_table.R
\name{toGt}
\alias{toGt}
\title{Gtsummary to Gt.}
\usage{
toGt(
  x,
  title = NULL,
  subtitle = NULL,
  opts = list(),
  subscripts = FALSE,
  fmt_markdown = FALSE
)
}
\arguments{
\item{x}{gtsummary table}

\item{title}{table title}

\item{subtitle}{table subtitle}

\item{opts}{gt tab options}

\item{subscripts}{use subscripts}

\item{fmt_markdown}{transform any markdown-formatted text, logical value. Default is FALSE.}
}
\description{
Gtsummary to Gt.
}
