% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_metrics_table.R
\name{toGt}
\alias{toGt}
\title{Gtsummary to Gt.}
\usage{
toGt(x, subscripts = FALSE, fmt_markdown = FALSE, opts = list())
}
\arguments{
\item{x}{gtsummary table}

\item{subscripts}{use subscripts}

\item{fmt_markdown}{transform any markdown-formatted text, logical value. Default is FALSE.}

\item{opts}{gt tab options}
}
\description{
Gtsummary to Gt.
}
