% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_trap_rule.R
\name{trap}
\alias{trap}
\title{NOTE: This method has been copied from the qpNCA package (qPharmetra).
 
Calculate Area Under the Curve Using Trapezoids.}
\usage{
trap(x, y, method = 1)
}
\arguments{
\item{x}{x variable, i.e. time}

\item{y}{y variable, i.e. concentration}

\item{method}{method:
* 1: linear up - linear down
* 2: linear up - logarithmic down
* 3: linear before Tmax, logarithmic after Tmax}
}
\value{
area (length-one numeric)
}
\description{
Calculates AUC using the trapezoidal method. Assumes data represent a single profile.
Despite choice of method, only linear interpolation is used
for areas of intervals beginning or ending with y: 0.
}
